+++ src/backends/FtpBackend.h
@@ -127,7 +127,7 @@ public slots:
     void deleteEntry(const QString &path, bool isDirectory) override;
     void renameEntry(const QString &oldPath, const QString &newPath) override;
     void moveEntry(const QString &oldPath, const QString &newPath, int requestId) override;
-    void createDirectory(const QString &path) override;
+    void createDirectory(const QString &path, bool ignoreAlreadyExists = false) override;
     void createFile(const QString &path) override;
     void setPermissions(const QString &path, int mode) override;